\name{Divers Functions}
\alias{divers}
\alias{asremlGinv}
\title{
  Divers functions and wrappers for GBLUP 
}
\description{
  \command{asreml} returns the inverse of the genomic
  relationship matrix in the asreml format.
}
\usage{
asremlGinv(G,variable='ID',rown=NULL)}
\arguments{
  \item{G}{SNP data}
  \item{variable,rown}{-- to do --} 
  \item{}{}
}


\value{
  \command{asreml} returns the inverse of the genomic
  relationship matrix in the asreml format.
 

}


\author{Malena Erbe, \email{malena.erbe@agr.uni-goettingen.de}
}

\examples{
## to do
}

\keyword{htest}
\keyword{ts}
